Update link to your `.zshrc` dot config
This is a digital hug from across the sea. You are a bad ass and I almost lost it when I went to install this plugin and saw your name. Where the heck do you find the time to maintain SO MANY lines of code?! Thanks again for everything you do. You are making a massive impact in the world. 

-Very Respectfully
Matt Kelly